Add files via upload
iNeuron-E-Commerce-Dashboard
Problem Statement
An online e-commerce company's analytics team wants to create a sales dashboard to evaluate sales based on different product categories. The business aims to provide people more choice over product categories so they may choose one and can observe the trend month- and product-wise as appropriate
Dataset
The dataset of the e-commerce company is provided in the form of Excel workbook. 
 The dataset contains the following columns: Order ID, Order Date, Ship Date, Aging, Ship Mode, Product Category, Product, Sales, Quantity, Discount, Profit, Shipping Cost, Order Priority, Customer ID, Customer Name, Segment, City, State, Country, Region, Months
Tool Used
Microsoft Excel is used to design the dashboard to gain insights about the sales and profits trends of the company.
Insights
 Electronics is the least-selling category and is preferred by age group 0-1 only. 
 The Central region generates the maximum sales for the company. 
 In categories other than the Electronics, the distribution of orders is almost similar for all the age groups. 
 The sales and profit trends for each category keeps on changing monthly.